+++ iot-ose-software-sensor/src/main/java/com/redhat/demo/iot/sensor/VibrationSensor.java
@@ -35,7 +35,7 @@ public class VibrationSensor implements Sensor {
 	@Value("${sensor.vibration.maxRange}") 
 	private int maxRange;
 	
-	public int currentValue;
+	public double currentValue;
 	
 	public int count = 0;
 	
@@ -68,7 +68,7 @@ public class VibrationSensor implements Sensor {
 		if(count > 0) {
 			
 			// Calculate random value from range
-			int randValue = ThreadLocalRandom.current().nextInt(minIteration, (maxIteration+1));
+			double randValue = ThreadLocalRandom.current().nextDouble(minIteration, (maxIteration+1));
 			currentValue = currentValue + randValue;
 			
 			if(currentValue < minRange || currentValue > maxRange) {
